Add support for checksums to http-tar and http-zip methods.
The compute-checksum function is the same for these methods as
the plain http method.

The SILENT parameter of url-retrieve-synchronously was set to true
because for some reason it was causing the text "GNUTLS peer has
closed connection" to be appended to the end of zip files. This
of course would cause checksums to fail.